+++ packages/vector_graphics/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.0.1
+  * Added `VectorGraphic` which consumes encoded vector graphics assets using
+    a `BytesLoader`.
+  * Added `AssetBytesLoader` and `NetworkBytesLoader` as example loader
+    implementations.
+
 ## 0.0.0
 
 * Create repository